update profiling script to handle comparison given an api reference url
