+++ scripts/performance_profiling_config.yaml
@@ -1,13 +1,16 @@
 params:
   # The backend API url of the deployment you wish to profile.
-  API_URL: "https://api.cellxgene.staging.single-cell.czi.technology"
+  API_URL: "https://pr-6420-backend.rdev.single-cell.czi.technology"
+
+  # OPTIONAL: The backend API url of the reference deployment you wish to profile against.
+  API_URL_REFERENCE: "https://pr-6427-backend.rdev.single-cell.czi.technology"
 
   # If you are profiling an rdev deployment, you must provide the auth cookie.
   # To get the auth cookie, log into the rdev deployment, open the browser dev tools, and copy the cookie from the request headers.
-  RDEV_AUTH_COOKIE: "_oauth2_proxy=..."
+  RDEV_AUTH_COOKIE: "_oauth2_proxy=l6seArXa8XPUStNeP168qXHqr5wCK2sjs9ZOPZTbv1Ih-JheEazWNxEH0Jc40UeGAxs3xP06ciFcmF3hGkTO0peMTa0jMzAmZggBnEeZKNkh5jsEuCRFIFri6GRc8-hmnQ9sZrj7vcS8MFvRokm8JW-5ImzHnZbtxTJHBdPOit30Cbw5PMW_R6KuMRXVXiDds5DwTDOIYl8UEGcT9gK-1hOzda5u1kGOOAI0tBtHvzLCF8oclmz628eKMrAF4bWlszxJV2XnpcvMQi1leGf_h29PXulgYM67P2JGsEEjT9QtIPL0q6MvgLJv5xPLU43SgnxR4mMcdSAI5LN_PX21aKXGrdsTqXuBOIoIDKdqi-cX_0ABHXOefzVP9s57ttknA3i6uISeZso1q25w1W1u0SwBwuLtrFdgh7yQmJHgkF1qsvHw1x1zpdF9KqWkEE3hGHnW05pohtPv0T0bNjrFjQ9f7ZpbK7ND2BPG5tZU7LAUZUgVP7I6Gfs1KS4aEnafF87IY_rII22EPFYBOM7lyabQSDljwmE2DXYUWo8Q-84MAc-T09h8ELzBLuxywBsJ3EWPsA8YjiYs5MDkmSQmdwsgWkWIAj3Zi2PhQoBMJPOzJXEUw8weKM0tgp0anR8EAT9n4AIuX92okYHXiRyMcgLvIkNb4YVXG3dma4uLLtL-61bvUchQ2-vMD8pvr4k_R4D21T2AbXYkQi-BkfaUXpdWirTPY3UB1ZJrLcOQotRaVCMc2kDm0j_CmPsY5DGKxFwDYD5GLB0vWKfF-Bxryyrg5Vsb-nRTOwUssHCdO4CG186TvbTiZvt7Yyd2FOLI9KaMh1nq1KLdd2E1g5qsxFPr_V1oIPPWaSXhCAkkmKcm65N_JvHaCJ7Xt_Y8KNxe6WbKtnSeytE28nh_YsWITq6sMpz9JrUmqOoGdHPdj3QyPiCpBKsPHvZn8RnxbWR_48rPvU8GNHz0j0VV163zqVdGFNz_95NajkaIEMsMmCNbZMLXUaOMMvxw345vE1EXdfgWoMjMOGLlS6IbEPtvcyppiSxgrqGp514gvfnBKzLvtkPQL5QexjkbvJFJK329SMZ2Xe_Y37be9vNSlAbGPaTbsvYvzstzIzpsQNVzQwzEZtI3i5ScOBmeSP3a5j0UW3vjGfasApdBYdyDzwgbA2pppdrG5vnWxfNF7YZFFmTm91bpRmtsZabEXSdvfupXqMikl9tnb-gIftTjdXVK5w-tLrYHoU0FmKfQ4F2Ic82s_Mrv4BEet3CPTEJhlmuiBdOvfaqt_uGGfxSfvxgal3sD3tJO-kW0oWhTIivVpXD4nimOeQj-VnofOmyGbOyJTvfX8UOd4-eUK7OfUDIUHIlLEkbTnWz_ENxbwPIUz68Cp5tskMjfbIorFxeEdg5ovO3RpElLLzojyUQygJxOIqcGL0UJjgENF1HDf2eByczPlXxKiBXOYIBKpUl_flooaxIrDJYEbVCGio0pAZvsT89KjlUJpYSYdMyjc-y7-f982I36Xv7nn139q4OUs62J_m8AJdpufKkU6FXMCHd_CWM3qpsLcGvYXnOcqjTrNQHlqugBogus3G3biLwGpP_6tkytiUo9xeUCdcbpQcUME_zMQHT12TIXgoWfCO2a-wMiZXZ7mwRmzvqwbHU7wP03xxg5gpAR2kTll23GwJCliRuLpIRULL76QJLB6J-wW8cLc_mT_2xz9uxgzH4D-OaPA_xesYkjuy_4Z4EIeBPRWDesdZAAvcgDmmaS0BLRUZ_iEfjFX4Qx4AbgvYre0-WTg0CsiqLdamcuA4E5X4Mhxh6H0ytvjNN-MksS2RPJJLJbwbmifvzqM2N6Lqa5wVekC6dZedY2-LZD2n_SZaLDqqJcaXbk_Zga14_IWjjxS7S8cJo_KpkhLLZIb2jFTUtb5VfQZghwj7oMxNiZeaJfeHzrb07scgvenHbBnSVkvWBBj31Xf_y8FVGCfMWMnebpM0iCM0rivO-F1xEaPE_4CcJ8agYFFKATdlsdATm8592HEoywLS9g1GS6xdv5enBZAvD50DLzQq44AMWyFrvC70hKmLiGO624R926GcyLSTmN0K3DS82dzHU3QzMj9Bfjd3bVbboFi9n7xlp2JUEM28Kg3jujloYNNxn3guiJXuJd8KCEiSKE5qjxzXm3uv5mzNs9qCkrIYi295YzELG3qci66fT2wGRnP4Z8CnVIYWKOcfJJ69z5PxKCZeSfC6vfIT4oXv-T4lkrlZmkV2VMLlMm7Q==|1704353689|2dCfPWZ9VpOa1du5lmxSZ_YiXhAJLudnv6dJGz08Mqo="
 
   # Outputs verbose logs if true.
   VERBOSE: false
 
   # Outputs summary plots of the profiling results if true. Requires matplotlib.
-  PLOTS: true
+  PLOTS: false
