+++ scripts/wmg_performance_profiling.py
@@ -1,3 +1,4 @@
+import ast
 import datetime
 import json
 import os
@@ -6,6 +7,69 @@ import sys
 import requests
 import yaml
 from tqdm import tqdm
+
+
+def generate_scatter_plots(json1, json2):
+    import matplotlib.pyplot as plt
+
+    # Load the profiling results
+    with open(json1, "r") as f1, open(json2, "r") as f2:
+        metrics = json.load(f1)["profiling_metrics_keys"]
+        profiling_results1 = json.load(f1)["profiles"]
+        profiling_results2 = json.load(f2)["profiles"]
+
+    # Create a new figure for each metric
+    compare_options = {d["params"]["compare"] for d in profiling_results1}
+    sex_criteria = [
+        ast.literal_eval(i)
+        for i in {str(d["params"]["payload"]["filter"]["sex_ontology_term_ids"]) for d in profiling_results1}
+    ]
+    num_conditions = len(compare_options) * len(sex_criteria)
+
+    fig, axs = plt.subplots(len(metrics), num_conditions)
+    fig.set_size_inches(4 * num_conditions, 12)
+
+    # Generate scatter plots for the collected metrics
+    for i, metric in enumerate(metrics):
+        for j, compare_option in enumerate(compare_options):
+            for k, sex in enumerate(sex_criteria):
+                metric_values1 = [
+                    d[metric]
+                    for d in profiling_results1
+                    if d["params"]["compare"] == compare_option
+                    and d["params"]["payload"]["filter"]["sex_ontology_term_ids"] == sex
+                ]
+                metric_values2 = [
+                    d[metric]
+                    for d in profiling_results2
+                    if d["params"]["compare"] == compare_option
+                    and d["params"]["payload"]["filter"]["sex_ontology_term_ids"] == sex
+                ]
+
+                max_value = max(metric_values1 + metric_values2)
+                row, col = i, j * len(sex_criteria) + k
+                axs[row, col].plot([0, max_value], [0, max_value], "k:")
+                axs[row, col].scatter(metric_values1, metric_values2)
+                axs[row, col].set_title(f"compare: {compare_option}, sex: {sex}")
+                axs[row, col].set_xlabel("Run 1")
+                axs[row, col].set_ylabel("Run 2")
+                axs[row, col].set_xlim(0, max_value + 0.5)
+                axs[row, col].set_ylim(0, max_value + 0.5)
+
+    # Manually add super-titles for each row
+    for i, metric in enumerate(metrics):
+        # Calculate y position for the super-title of each row
+        y_position = 1 - (i - 0.01) / len(metrics)
+        fig.text(0.5, y_position, metric, ha="center", va="center", fontsize=14, weight="bold")
+
+    fig.tight_layout(pad=3.0)
+    plt.rcParams.update({"font.size": 10})
+
+    # Save the scatter plots to a PDF
+    timestamp = datetime.datetime.now().strftime("%Y%m%d_%H%M%S")
+    fig.savefig(f"scatter_plots_{timestamp}.pdf")
+    print(f"Scatter plots output to scatter_plots_{timestamp}.pdf")
+
 
 # Add the root directory to the Python module search path so you can reference backend
 # without needing to move this script to the root directory to run it.
@@ -448,107 +512,120 @@ if __name__ == "__main__":
     # Parsing parameters from the config file
     rdev_auth_cookie = data["params"]["RDEV_AUTH_COOKIE"]
     api_url = data["params"]["API_URL"]
+    api_url_reference = data["params"]["API_URL_REFERENCE"]
     verbose = data["params"]["VERBOSE"]
     generate_plots = data["params"]["PLOTS"]
 
-    profiling_dicts = []
-    profiling_metrics_keys = []
-    for i, body in enumerate(tqdm(POST_BODIES, disable=verbose)):
-        if verbose:
-            print(f"Executing query {i+1}/{len(POST_BODIES)} with body: {body}")
+    api_urls = [api_url]
+    if api_url_reference != "":
+        api_urls.append(api_url_reference)
 
-        headers = {"Content-Type": "application/json"}
-        if rdev_auth_cookie != "_oauth2_proxy=..." and rdev_auth_cookie != "":
-            headers["Cookie"] = rdev_auth_cookie
-        response = requests.post(f"{api_url}/wmg/v2/query", data=json.dumps(body), headers=headers)
+    output_files = []
+    for api_url in api_urls:
+        print("Profiling API:", api_url)
 
-        server_timing_header = response.headers.get("Server-Timing")
-        assert response.status_code == 200
-        assert server_timing_header is not None
+        profiling_dicts = []
+        profiling_metrics_keys = []
+        for i, body in enumerate(tqdm(POST_BODIES, disable=verbose)):
+            if verbose:
+                print(f"Executing query {i+1}/{len(POST_BODIES)} with body: {body}")
 
-        # Parse the server timing header to get the profiling results
-        server_timing_values = server_timing_header.split(", ")
-        profiling_dict = {}
-        total_time = 0
-        for value in server_timing_values:
-            name, dur = value.split(";dur=")
-            dur, _ = dur.split(";desc")
-            dur = float(dur) / 1000
-            profiling_dict[name] = dur
-            total_time += dur
-            profiling_metrics_keys.append(name)
-        response_time = response.elapsed.total_seconds()
-        download_time = response_time - total_time
-        profiling_dict["total_time_backend"] = total_time
-        profiling_dict["total_time_response"] = response_time
-        profiling_dict["download_time"] = download_time
-        response_size = len(response.content) / (1024 * 1024)
-        profiling_dict["response_size_mb"] = response_size
+            headers = {"Content-Type": "application/json"}
+            if rdev_auth_cookie != "_oauth2_proxy=..." and rdev_auth_cookie != "":
+                headers["Cookie"] = rdev_auth_cookie
+            response = requests.post(f"{api_url}/wmg/v2/query", data=json.dumps(body), headers=headers)
 
-        if verbose:
-            print("Profiling results:", profiling_dict)
+            server_timing_header = response.headers.get("Server-Timing")
+            assert response.status_code == 200
+            assert server_timing_header is not None
 
-        profiling_dict["params"] = {
-            "payload": body,
-            "num_genes": len(body["filter"]["gene_ontology_term_ids"]),
-            "compare": body.get("compare"),
+            # Parse the server timing header to get the profiling results
+            server_timing_values = server_timing_header.split(", ")
+            profiling_dict = {}
+            total_time = 0
+            for value in server_timing_values:
+                name, dur = value.split(";dur=")
+                dur, _ = dur.split(";desc")
+                dur = float(dur) / 1000
+                profiling_dict[name] = dur
+                total_time += dur
+                profiling_metrics_keys.append(name)
+            response_time = response.elapsed.total_seconds()
+            download_time = response_time - total_time
+            profiling_dict["total_time_backend"] = total_time
+            profiling_dict["total_time_response"] = response_time
+            profiling_dict["download_time"] = download_time
+            response_size = len(response.content) / (1024 * 1024)
+            profiling_dict["response_size_mb"] = response_size
+
+            if verbose:
+                print("Profiling results:", profiling_dict)
+
+            profiling_dict["params"] = {
+                "payload": body,
+                "num_genes": len(body["filter"]["gene_ontology_term_ids"]),
+                "compare": body.get("compare"),
+            }
+            profiling_dicts.append(profiling_dict)
+
+        profiling_results = {
+            "profiles": profiling_dicts,
+            "api_url": api_url,
+            "total_time_backend": sum([d["total_time_backend"] for d in profiling_dicts]),
+            "total_time_response": sum([d["total_time_response"] for d in profiling_dicts]),
+            "total_time_download": sum([d["download_time"] for d in profiling_dicts]),
+            "total_response_size_mb": sum([d["response_size_mb"] for d in profiling_dicts]),
+            "profiling_metrics_keys": profiling_metrics_keys,
         }
-        profiling_dicts.append(profiling_dict)
+        print(f"Total number of queries executed: {len(profiling_dicts)}")
+        print("Total time backend:", profiling_results["total_time_backend"])
+        print("Total time response:", profiling_results["total_time_response"])
+        print("Total time download:", profiling_results["total_time_download"])
+        print("Total response size (MB):", profiling_results["total_response_size_mb"])
 
-    profiling_results = {
-        "profiles": profiling_dicts,
-        "api_url": api_url,
-        "total_time_backend": sum([d["total_time_backend"] for d in profiling_dicts]),
-        "total_time_response": sum([d["total_time_response"] for d in profiling_dicts]),
-        "total_time_download": sum([d["download_time"] for d in profiling_dicts]),
-        "total_response_size_mb": sum([d["response_size_mb"] for d in profiling_dicts]),
-        "profiling_metrics_keys": profiling_metrics_keys,
-    }
-    print(f"Total number of queries executed: {len(profiling_dicts)}")
-    print("Total time backend:", profiling_results["total_time_backend"])
-    print("Total time response:", profiling_results["total_time_response"])
-    print("Total time download:", profiling_results["total_time_download"])
-    print("Total response size (MB):", profiling_results["total_response_size_mb"])
+        # Write the profiling results to a file
+        timestamp = datetime.datetime.now().strftime("%Y%m%d_%H%M%S")
+        with open(f"profiling_results_{timestamp}.json", "w") as f:
+            json.dump(profiling_results, f)
 
-    # Write the profiling results to a file
-    timestamp = datetime.datetime.now().strftime("%Y%m%d_%H%M%S")
-    with open(f"profiling_results_{timestamp}.json", "w") as f:
-        json.dump(profiling_results, f)
+            print(f"Profiling results output to {f.name}")
+            output_files.append(f.name)
 
-        print(f"Profiling results output to {f.name}")
+        # Create a new figure for each metric
+        if generate_plots:
+            import matplotlib.pyplot as plt
 
-    # Create a new figure for each metric
-    if generate_plots:
-        import matplotlib.pyplot as plt
+            # Generate matplotlib plots for the collected metrics
+            metrics = profiling_metrics_keys + ["total_time_backend"]
+            num_genes = [d["params"]["num_genes"] for d in profiling_dicts]
+            compare_options = {d["params"]["compare"] for d in profiling_dicts}
 
-        # Generate matplotlib plots for the collected metrics
-        metrics = profiling_metrics_keys + ["total_time_backend"]
-        num_genes = [d["params"]["num_genes"] for d in profiling_dicts]
-        compare_options = {d["params"]["compare"] for d in profiling_dicts}
+            fig, axs = plt.subplots(len(metrics), 1)
+            fig.set_size_inches(8, 12)
+            for i, metric in enumerate(metrics):
+                for compare_option in compare_options:
+                    for sex_criteria in SEX_CRITERIA:
+                        compare_dicts = [
+                            d
+                            for d in profiling_dicts
+                            if d["params"]["compare"] == compare_option
+                            and d["params"]["payload"]["filter"]["sex_ontology_term_ids"] == sex_criteria
+                        ]
+                        metric_values = [d[metric] for d in compare_dicts]
+                        num_genes_compare = [d["params"]["num_genes"] for d in compare_dicts]
 
-        fig, axs = plt.subplots(len(metrics), 1)
-        fig.set_size_inches(8, 12)
-        for i, metric in enumerate(metrics):
-            for compare_option in compare_options:
-                for sex_criteria in SEX_CRITERIA:
-                    compare_dicts = [
-                        d
-                        for d in profiling_dicts
-                        if d["params"]["compare"] == compare_option
-                        and d["params"]["payload"]["filter"]["sex_ontology_term_ids"] == sex_criteria
-                    ]
-                    metric_values = [d[metric] for d in compare_dicts]
-                    num_genes_compare = [d["params"]["num_genes"] for d in compare_dicts]
+                        axs[i].plot(num_genes_compare, metric_values, label=f"{compare_option} {sex_criteria}")
+                        axs[i].set_title(f"{metric} vs number of genes")
+                        axs[i].set_xlabel("number of genes")
+                        axs[i].set_ylabel(metric)
+                axs[i].legend(loc="center left", bbox_to_anchor=(1, 0.5))
+            fig.subplots_adjust(right=0.85)  # Adjust the whitespace on the right
+            fig.tight_layout(pad=3.0)
+            plt.rcParams.update({"font.size": 10})
 
-                    axs[i].plot(num_genes_compare, metric_values, label=f"{compare_option} {sex_criteria}")
-                    axs[i].set_title(f"{metric} vs number of genes")
-                    axs[i].set_xlabel("number of genes")
-                    axs[i].set_ylabel(metric)
-            axs[i].legend(loc="center left", bbox_to_anchor=(1, 0.5))
-        fig.subplots_adjust(right=0.85)  # Adjust the whitespace on the right
-        fig.tight_layout(pad=3.0)
-        plt.rcParams.update({"font.size": 10})
+            # Save the plots to a PDF
+            fig.savefig(f"profiling_plots_{timestamp}.pdf")
+            print(f"Profiling plots output to profiling_plots_{timestamp}.pdf")
 
-        # Save the plots to a PDF
-        fig.savefig(f"profiling_plots_{timestamp}.pdf")
-        print(f"Profiling plots output to profiling_plots_{timestamp}.pdf")
+    if len(api_urls) == 2:
+        generate_scatter_plots(output_files[0], output_files[1])
